Consolidate ztf models and rename methods.
Method renaming to favor namespace usage, e.g.:

  From: services.query_ztf_data.query_ztf_found_data
  To: services.query_ztf_data.found
